+++ store/v2/root/factory.go
@@ -14,9 +14,7 @@ import (
 	"go.cosmonity.xyz/evolve/store/v2/db"
 	"go.cosmonity.xyz/evolve/store/v2/internal"
 	"go.cosmonity.xyz/evolve/store/v2/metrics"
-	"go.cosmonity.xyz/evolve/store/v2/migration"
 	"go.cosmonity.xyz/evolve/store/v2/pruning"
-	"go.cosmonity.xyz/evolve/store/v2/snapshots"
 
 	"cosmossdk.io/core/log"
 	corestore "cosmossdk.io/core/store"
@@ -116,30 +114,14 @@ func CreateRootStore(opts *FactoryOptions) (store.RootStore, error) {
 	}
 
 	// check if we need to migrate the store
-	isMigrating := false
-	scType := storeOpts.SCType
 
-	if scType != SCTypeIavl {
-		isMigrating = true  // need to migrate
-		scType = SCTypeIavl // only support iavl v1 for migration
-	}
+	scType := storeOpts.SCType
 
 	trees := make(map[string]commitment.Tree, len(opts.StoreKeys))
 	for _, key := range opts.StoreKeys {
 		tree, err := newTreeFn(key, scType)
 		if err != nil {
 			return nil, err
-		}
-		if isMigrating {
-			v, err := tree.GetLatestVersion()
-			if err != nil {
-				return nil, err
-			}
-			if v == 0 && latestVersion > 0 {
-				if err := tree.SetInitialVersion(latestVersion + 1); err != nil {
-					return nil, err
-				}
-			}
 		}
 		trees[key] = tree
 	}
@@ -157,31 +139,6 @@ func CreateRootStore(opts *FactoryOptions) (store.RootStore, error) {
 		return nil, err
 	}
 
-	var mm *migration.Manager
-	if isMigrating {
-		snapshotDB, err := snapshots.NewStore(fmt.Sprintf("%s/data/snapshots/store.db", opts.RootDir))
-		if err != nil {
-			return nil, err
-		}
-		snapshotMgr := snapshots.NewManager(snapshotDB, snapshots.SnapshotOptions{}, sc, nil, opts.Logger)
-		var newSC *commitment.CommitStore
-		if scType != storeOpts.SCType {
-			newTrees := make(map[string]commitment.Tree, len(opts.StoreKeys))
-			for _, key := range opts.StoreKeys {
-				tree, err := newTreeFn(key, storeOpts.SCType)
-				if err != nil {
-					return nil, err
-				}
-				newTrees[key] = tree
-			}
-			newSC, err = commitment.NewCommitStore(newTrees, nil, opts.SCRawDB, opts.Logger)
-			if err != nil {
-				return nil, err
-			}
-		}
-		mm = migration.NewManager(opts.SCRawDB, snapshotMgr, newSC, opts.Logger)
-	}
-
 	pm := pruning.NewManager(sc, storeOpts.SCPruningOption)
-	return New(opts.SCRawDB, opts.Logger, sc, pm, mm, metrics.NoOpMetrics{})
+	return New(opts.SCRawDB, opts.Logger, sc, pm, metrics.NoOpMetrics{})
 }