+++ store/v2/root/store_test.go
@@ -53,7 +53,7 @@ func (s *RootStoreTestSuite) SetupTest() {
 	s.Require().NoError(err)
 
 	pm := pruning.NewManager(sc, nil)
-	rs, err := New(dbm.NewMemDB(), noopLog, sc, pm, nil, nil)
+	rs, err := New(dbm.NewMemDB(), noopLog, sc, pm, nil)
 	s.Require().NoError(err)
 
 	s.rootStore = rs
@@ -74,7 +74,7 @@ func (s *RootStoreTestSuite) newStoreWithPruneConfig(config *store.PruningOption
 
 	pm := pruning.NewManager(sc, config)
 
-	rs, err := New(dbm.NewMemDB(), noopLog, sc, pm, nil, nil)
+	rs, err := New(dbm.NewMemDB(), noopLog, sc, pm, nil)
 	s.Require().NoError(err)
 
 	s.rootStore = rs
@@ -83,7 +83,7 @@ func (s *RootStoreTestSuite) newStoreWithPruneConfig(config *store.PruningOption
 func (s *RootStoreTestSuite) newStoreWithBackendMount(sc store.Committer, pm *pruning.Manager) {
 	noopLog := coretesting.NewNopLogger()
 
-	rs, err := New(dbm.NewMemDB(), noopLog, sc, pm, nil, nil)
+	rs, err := New(dbm.NewMemDB(), noopLog, sc, pm, nil)
 	s.Require().NoError(err)
 
 	s.rootStore = rs